==> R/02_D_GeoMarkers.Rmd <==
---
title: "Mapping geographical markers in top Ecology papers"
author: "Gabriel Nakamura"
date: "2023-03-06"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Reading required data and libraries

All libraries needed to run the analysis 

```{r}
library(here)
library(sf)
library(dplyr)
library(tidyverse)
library(magrittr)
library(bibliometrix)
library(rnaturalearth)
library(ggplot2)
library(rmapshaper)
```

Reading all files containing the top 1000 articles selected based on the address location of the corresponding author

```{r}
files_LA <- here::here("data", "raw", "LA_top1000.txt")
M_LA <- convert2df(file = files_LA, dbsource = 'wos', format = "plaintext")
files_US <- here::here("data", "raw", "USA_CAN_top1000.txt")
M_US <- convert2df(file = files_US, dbsource = 'wos', format = "plaintext")
files_EU <- here::here("data", "raw", "EU_top1000.txt")
M_EU <- convert2df(file = files_EU, dbsource = 'wos', format = "plaintext")
files_EA <- here::here("data", "raw", "EastAsia_top1000.txt")
M_EA <- convert2df(file = files_EA, dbsource = 'wos', format = "plaintext")
files_SA <- here::here("data", "raw", "SouthAsia_top1000.txt")
M_SA <- convert2df(file = files_SA, dbsource = 'wos', format = "plaintext")
files_sahara <- here::here("data", "raw", "SubSahara_top1000.txt")
M_sahara <- convert2df(file = files_sahara, dbsource = 'wos', format = "plaintext")
files_mafrica <- here::here("data", "raw", "MiddleEastAfrica_top1000.txt")
M_mafrica <- convert2df(file = files_mafrica, dbsource = 'wos', format = "plaintext")

```

# Calculating the number of geo markers

Here I calculate the number of [geographical markers](https://www.sciencedirect.com/science/article/abs/pii/S0016718519301551) per country for each one of the groups defined in the previous section.

First I defined the countries in each political region based on the classification contained in {`rnaturalearth`} package. This classification is based on the classification made by the World Bank (column `region_wb`). I selected all the names of the countries in each region and separate spatial objects for each

```{r}
countries <- ne_countries(scale = "medium", returnclass = "sf")
names_LA <- subset(countries, region_wb == "Latin America & Caribbean")$admin
sf_la <- subset(countries, region_wb == "Latin America & Caribbean")[, c("admin", "region_wb", "geometry")]
names_NA <- subset(countries, region_wb == "North America")$admin
names_NA <- c(names_NA, "USA", "Usa", "United States")
sf_na <- subset(countries, region_wb == "North America")[, c("admin", "region_wb", "geometry")]
names_EU <- subset(countries, region_wb == "Europe & Central Asia")$admin
names_EU <- c(names_EU, "England", "Scotland", "Wales", "UK")
sf_eu <- subset(countries, region_wb == "Europe & Central Asia")[, c("admin", "region_wb", "geometry")]
names_MidleEastAfrica <- subset(countries, region_wb == "Middle East & North Africa")$admin
sf_africa <- subset(countries, region_wb == "Middle East & North Africa")[, c("admin", "region_wb", "geometry")]
names_EastAsia <- subset(countries, region_wb == "East Asia & Pacific")$admin
sf_asia <- subset(countries, region_wb == "East Asia & Pacific")[, c("admin", "region_wb", "geometry")]
names_SouthAsia <- subset(countries, region_wb == "South Asia")$admin
sf_southasia <- subset(countries, region_wb == "South Asia")[, c("admin", "region_wb", "geometry")]
names_SubSahara <- subset(countries, region_wb == "Sub-Saharan Africa")$admin
sf_SubSahara <- subset(countries, region_wb == "Sub-Saharan Africa")[, c("admin", "region_wb", "geometry")]
```

I've made some minor manual complements to add names that are from some region but were not present in the rnaturalearth database

# Calculating the number of zoogeographical markers

Here I calculated the number of geographical markers for each country in each region for all top 1000 cited papers.

```{r}

# Zoogeographical regions

zoogeo <- c("Neotropical", "Neotropic", "Neartical", "Neartic", "Afrotropic", "Afrotropical", "Paleartical", "Paleartic", "Australasian", "Indomalayan")


# Europe
geo_eu <- lapply(toupper(names_EU), function(x) grep(x, M_EU$TI, fixed = T))
zoo_eu <- lapply(toupper(zoogeo), function(x) grep(x, M_EU$TI, fixed = T))
count_geo_eu <- lapply(geo_eu, function(x) ifelse(length(x) == 0, 0, length(x)))
count_zoo_eu <- lapply(zoo_eu, function(x) ifelse(length(x) == 0, 0, length(x)))
names(count_geo_eu) <- names_EU
names(count_zoo_eu) <- zoogeo
count_geo_eu <- unlist(count_geo_eu)
count_zoo_eu <- unlist(count_zoo_eu)

# Latin America

geo_la <- lapply(toupper(names_LA), function(x) grep(x, M_LA$TI, fixed = T))
zoo_la <- lapply(toupper(zoogeo), function(x) grep(x, M_LA$TI, fixed = T))
count_geo_la <- lapply(geo_la, function(x) ifelse(length(x) == 0, 0, length(x)))
count_zoo_la <- lapply(zoo_la, function(x) ifelse(length(x) == 0, 0, length(x)))
names(count_geo_la) <- names_LA
names(count_zoo_la) <- zoogeo
count_geo_la <- unlist(count_geo_la)
count_zoo_la <- unlist(count_zoo_la)

# North America

geo_us <- lapply(toupper(names_NA), function(x) grep(x, M_US$TI, fixed = T))
zoo_us <- lapply(toupper(zoogeo), function(x) grep(x, M_US$TI, fixed = T))
count_geo_us <- lapply(geo_us, function(x) ifelse(length(x) == 0, 0, length(x)))
count_zoo_us <- lapply(zoo_us, function(x) ifelse(length(x) == 0, 0, length(x)))
names(count_geo_us) <- names_NA
names(count_zoo_us) <- zoogeo
count_geo_us <- unlist(count_geo_us)
count_zoo_us <- unlist(count_zoo_us)

# East Asia

geo_ea <- lapply(toupper(names_EastAsia), function(x) grep(x, M_EA$TI, fixed = T))
zoo_ea <- lapply(toupper(zoogeo), function(x) grep(x, M_EA$TI, fixed = T))
count_geo_ea <- lapply(geo_ea, function(x) ifelse(length(x) == 0, 0, length(x)))
count_zoo_ea <- lapply(zoo_ea, function(x) ifelse(length(x) == 0, 0, length(x)))
names(count_geo_ea) <- names_EastAsia
names(count_zoo_ea) <- zoogeo
count_geo_ea <- unlist(count_geo_ea)
count_zoo_ea <- unlist(count_zoo_ea)

# South Asia

geo_sa <- lapply(toupper(names_SouthAsia), function(x) grep(x, M_SA$TI, fixed = T))
zoo_sa <- lapply(toupper(zoogeo), function(x) grep(x, M_SA$TI, fixed = T))
count_geo_sa <- lapply(geo_sa, function(x) ifelse(length(x) == 0, 0, length(x)))
count_zoo_sa <- lapply(zoo_sa, function(x) ifelse(length(x) == 0, 0, length(x)))
names(count_geo_sa) <- names_SouthAsia
names(count_zoo_sa) <- zoogeo
count_geo_sa <- unlist(count_geo_sa)
count_zoo_sa <- unlist(count_zoo_sa)


# Africa 
geo_africa <- lapply(toupper(names_SubSahara), function(x) grep(x, M_sahara$TI, fixed = T))
zoo_africa <- lapply(toupper(zoogeo), function(x) grep(x, M_sahara$TI, fixed = T))
count_geo_africa <- lapply(geo_africa, function(x) ifelse(length(x) == 0, 0, length(x)))
count_zoo_africa <- lapply(zoo_africa, function(x) ifelse(length(x) == 0, 0, length(x)))
names(count_geo_africa) <- names_SubSahara
names(count_zoo_africa) <- zoogeo
count_geo_africa <- unlist(count_geo_africa)
count_zoo_africa <- unlist(count_zoo_africa)


# Middle North Africa
geo_NAfrica <- lapply(toupper(names_MidleEastAfrica), function(x) grep(x, M_mafrica$TI, fixed = T))
zoo_NAfrica <- lapply(toupper(zoogeo), function(x) grep(x, M_mafrica$TI, fixed = T))
count_geo_NAfrica <- lapply(geo_NAfrica, function(x) ifelse(length(x) == 0, 0, length(x)))
count_zoo_NAfrica <- lapply(zoo_NAfrica, function(x) ifelse(length(x) == 0, 0, length(x)))
names(count_geo_NAfrica) <-  names_MidleEastAfrica
names(count_zoo_NAfrica) <- zoogeo
count_geo_NAfrica <- unlist(count_geo_NAfrica)
count_zoo_NAfrica <- unlist(count_zoo_NAfrica)

df_zoo_count <- 
  rbind(count_zoo_africa, count_zoo_ea, count_zoo_eu, count_zoo_la, count_zoo_NAfrica, count_zoo_us)
colSums(df_zoo_count)
```



# Joining geographical markers count with spatial tables

```{r}

countries <- st_transform(countries, crs = "+proj=robin")

df_count_la <- data.frame(admin = names(count_geo_la), count = as.numeric(count_geo_la))
sf_count_la <- 
  sf_la %>% 
  right_join(df_count_la, by = "admin") %>% 
  st_transform(crs = "+proj=robin")

df_count_us <- data.frame(admin = names(count_geo_us), count = as.numeric(count_geo_us))
US <- sum(count_geo_us[5:7])
df_count_us[4, "count"] <- US
df_count_us1 <- df_count_us[-c(5:7), ] 
sf_count_us <- 
  sf_na %>% 
  right_join(df_count_us1, by = "admin") %>% 
  st_transform(crs = "+proj=robin")

df_count_asia <- data.frame(admin = names(count_geo_ea), count = as.numeric(count_geo_ea))
sf_count_ea <- 
  sf_asia %>% 
  right_join(df_count_asia, by = "admin") %>% 
  st_transform(crs = "+proj=robin")

df_count_southasia <- data.frame(admin = names(count_geo_sa), count = as.numeric(count_geo_sa))
sf_count_sa <- 
  sf_southasia %>% 
  right_join(df_count_southasia, by = "admin") %>% 
  st_transform(crs = "+proj=robin")


df_count_eu <- data.frame(admin = names(count_geo_eu), count = as.numeric(count_geo_eu))
UK <- c("England", "Scotland", "Wales", "UK")
UK_sum <- sum(df_count_eu[match(UK, df_count_eu$admin), "count"])
df_count_eu[which(df_count_eu$admin == "United Kingdom"), "count"] <- UK_sum
df_count_eu1 <- df_count_eu[-match(UK, df_count_eu$admin), ] 
sf_count_eu <- 
  sf_eu %>% 
  right_join(df_count_eu1, by = "admin") %>% 
  st_transform(crs = "+proj=robin")
  
df_count_africa <- data.frame(admin = names(count_geo_africa), count = as.numeric(count_geo_africa))
sf_count_africa <- 
  sf_SubSahara %>% 
  right_join(df_count_africa, by = "admin") %>% 
  st_transform(crs = "+proj=robin")


df_count_NAfrica <- data.frame(admin = names(count_geo_NAfrica), count = as.numeric(count_geo_NAfrica))
sf_count_NAfrica <- 
  sf_africa %>% 
  right_join(df_count_NAfrica, by = "admin") %>% 
  st_transform(crs = "+proj=robin")

sf_all <- rbind(sf_count_africa, sf_count_ea, sf_count_sa, sf_count_eu, sf_count_la, sf_count_NAfrica, sf_count_us)
sf_mean <- sf_all %>% group_by(region_wb) %>% 
  summarise(sum_count=sum(count),
            .groups = 'drop')

```


Making maps showing the number of geographical markers for each WB region

```{r}

map_markers_wb <-
  ggplot() +
  geom_sf(data = sf_mean, aes(fill = sum_count)) +
  rcartocolor::scale_fill_carto_c(palette = "DarkMint",
                                  direction = 1,
                                  limits = c(10, 300),  ## max percent overall
                                  breaks = seq(10, 300, by = 50),
                                  labels = glue::glue("{seq(10, 300, by = 50)}")) +
  guides(fill = guide_colorbar(barheight = unit(3, units = "mm"),
                               barwidth = unit(50, units = "mm"),
                               direction = "horizontal",
                               ticks.colour = "grey20",
                               title.position = "top",
                               label.position = "bottom",
                               title.hjust = 0.5)) +
  labs(fill = "Total Markers", subtitle = "") +
  xlab("") +
  ylab("") +
  theme(legend.position = "bottom",
        panel.background = element_rect(fill = "transparent"),
        plot.margin = unit(c(0.1, 0.1, 0.1, 0.1), "mm"),
        legend.text = element_text(family = "Times", color = "black", size = 10),
        axis.text = element_blank(),
        axis.ticks.x.bottom = element_blank(),
        panel.border = element_blank())
  

map_markers_wb


```
