==> R/02_D_rarefaction_citation.Rmd <==
---
title: "Citation Bias with rarefaction"
author: "Gabriel Nakamura"
date: "2023-06-07"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Rarefaction analysis with citation values for each region

Here I performed a rarefaction analysis with citation values for each region. This analysis is used to produce Figure 1 in the main text. The aim with this analysis is to reduce possible biases created by the variation in the number of citations that each region receive.

```{r}
library(hrbrthemes)
library(ggplot2)
library(bibliometrix)
library(cowplot)
library(rnaturalearth)


countries <- ne_countries(scale = "medium", returnclass = "sf")
names_LA <- subset(countries, region_wb == "Latin America & Caribbean")$admin
names_NA <- subset(countries, region_wb == "North America")$admin
names_NA <- c(names_NA, "USA", "Usa", "United States")
names_EU <- subset(countries, region_wb == "Europe & Central Asia")$admin
names_EU <- c(names_EU, "England", "Scotland", "Wales", "UK")
names_MidleEastAfrica <- subset(countries, region_wb == "Middle East & North Africa")$admin
names_EastAsia <- subset(countries, region_wb == "East Asia & Pacific")$admin
names_SouthAsia <- subset(countries, region_wb == "South Asia")$admin
names_SubSahara <- subset(countries, region_wb == "Sub-Saharan Africa")$admin

```

Rarefaction procedure: for each region (following World Bank) I sampled without replacement 1000 articles 100 times from matrix M (matrix containing articles that cited the top 1000 single author articles from each region)

```{r}
# reading M matrices ------------------------------------------------------

M_citations_EastAsia <- readRDS(here::here("data", "processed", "M_EastAsia.rds"))
M_citations_EU <- readRDS(here::here("data", "processed", "M_EU.rds"))
M_citations_la <- readRDS(here::here("data", "processed", "M_LA.rds"))
M_citations_MiddleAfrica <- readRDS(here::here("data", "processed", "M_MiddleAfrica.rds"))
M_citations_Sahara <- readRDS(here::here("data", "processed", "M_Sahara.rds"))
M_citations_usa_can <- readRDS(here::here("data", "processed", "M_USA.rds"))

# general setup of rarefaction procedure ------------------------------------

samp <- 1000 # number of samples
iter <- 100 # number of times to repeat sampling procedure
list_samp_eastasia <- lapply(seq_along(along.with = 1:100), function(x) sample(1:nrow(M_citations_EastAsia), size = samp, replace = FALSE))
list_samp_middleafrica <- lapply(seq_along(along.with = 1:100), function(x) sample(1:nrow(M_citations_MiddleAfrica), size = samp, replace = FALSE))
list_samp_eu <- lapply(seq_along(along.with = 1:100), function(x) sample(1:nrow(M_citations_EU), size = samp, replace = FALSE))
list_samp_la <- lapply(seq_along(along.with = 1:100), function(x) sample(1:nrow(M_citations_la), size = samp, replace = FALSE))
list_samp_sahara <- lapply(seq_along(along.with = 1:100), function(x) sample(1:nrow(M_citations_Sahara), size = samp, replace = FALSE))
list_samp_usacan <- lapply(seq_along(along.with = 1:100), function(x) sample(1:nrow(M_citations_usa_can), size = samp, replace = FALSE))


# sampling each region ----------------------------------------------------

samp_eastasia <- lapply(list_samp_eastasia, function(x) M_citations_EastAsia[x,])
samp_eu <- lapply(list_samp_eu, function(x) M_citations_EU[x,])
samp_la <- lapply(list_samp_la, function(x) M_citations_la[x,])
samp_middleafrica <- lapply(list_samp_middleafrica, function(x) M_citations_MiddleAfrica[x,])
samp_sahara <- lapply(list_samp_sahara, function(x) M_citations_Sahara[x,])
samp_usacan <- lapply(list_samp_usacan, function(x) M_citations_usa_can[x,])


# biblio analysis with all bootstrap matrices -----------------------------

res_eastasia <- lapply(samp_eastasia, function(x) biblioAnalysis(x, sep = ";")) 
res_eu <- lapply(samp_eu, function(x) biblioAnalysis(x, sep = ";")) 
res_la <- lapply(samp_la, function(x) biblioAnalysis(x, sep = ";")) 
res_middleafrica <- lapply(samp_middleafrica, function(x) biblioAnalysis(x, sep = ";")) 
res_sahara <- lapply(samp_sahara, function(x) biblioAnalysis(x, sep = ";")) 
res_usacan <- lapply(samp_usacan, function(x) biblioAnalysis(x, sep = ";")) 

saveRDS(res_eastasia, here::here("output", "res_boot_eastasia.rds"))
saveRDS(res_eu, here::here("output", "res_boot_eu.rds"))
saveRDS(res_la, here::here("output", "res_boot_la.rds"))
saveRDS(res_middleafrica, here::here("output", "res_boot_middleafrica.rds"))
saveRDS(res_sahara, here::here("output", "res_boot_sahara.rds"))
saveRDS(res_usacan, here::here("output", "res_boot_usacan.rds"))
```

Getting data from rarefaction analysis

```{r}
# summarizing results from bootstrap --------------------------------------

res_eastasia <- readRDS(here::here("output", "res_boot_eastasia.rds"))
res_eu <- readRDS(here::here("output", "res_boot_eu.rds"))
res_la <- readRDS(here::here("output", "res_boot_la.rds"))
res_middleafrica <- readRDS(here::here("output", "res_boot_middleafrica.rds"))
res_sahara <- readRDS(here::here("output", "res_boot_sahara.rds"))
res_usacan <- readRDS(here::here("output", "res_boot_usacan.rds"))


cit_eastasia <- lapply(res_eastasia, function(x) x$Countries)
long_cit_eastasia <- data.frame(count = names(unlist(cit_eastasia)), ncitations = unlist(cit_eastasia))
library(dplyr)
mean_eastasia <- 
long_cit_eastasia %>% 
  group_by(count) %>% 
  summarize(Mean = mean(ncitations, na.rm = T), std = sd(ncitations, na.rm = T))
cit_eastasia <- as.data.frame(mean_eastasia[order(mean_eastasia$Mean, decreasing = T), ])


cit_eu <- lapply(res_eu, function(x) x$Countries)
long_cit_eu <- data.frame(count = names(unlist(cit_eu)), ncitations = unlist(cit_eu))
mean_eu <- 
  long_cit_eu %>% 
  group_by(count) %>% 
  summarize(Mean = mean(ncitations, na.rm = T), std = sd(ncitations, na.rm = T))
cit_eu <- as.data.frame(mean_eu[order(mean_eu$Mean, decreasing = T), ])


cit_la <- lapply(res_la, function(x) x$Countries)
long_cit_la <- data.frame(count = names(unlist(cit_la)), ncitations = unlist(cit_la))
mean_la <- 
  long_cit_la %>% 
  group_by(count) %>% 
  summarize(Mean = mean(ncitations, na.rm = T), std = sd(ncitations, na.rm = T))
cit_la <- as.data.frame(mean_la[order(mean_la$Mean, decreasing = T), ])

cit_middleafrica <- lapply(res_middleafrica, function(x) x$Countries)
long_cit_middleafrica <- data.frame(count = names(unlist(cit_middleafrica)), ncitations = unlist(cit_middleafrica))
mean_middleafrica <- 
  long_cit_middleafrica %>% 
  group_by(count) %>% 
  summarize(Mean = mean(ncitations, na.rm = T), std = sd(ncitations, na.rm = T))
cit_middleafrica <- as.data.frame(mean_middleafrica[order(mean_middleafrica$Mean, decreasing = T), ])

cit_sahara <- lapply(res_sahara, function(x) x$Countries)
long_cit_sahara <- data.frame(count = names(unlist(cit_sahara)), ncitations = unlist(cit_sahara))
mean_sahara <- 
  long_cit_sahara %>% 
  group_by(count) %>% 
  summarize(Mean = mean(ncitations, na.rm = T), std = sd(ncitations, na.rm = T))
cit_sahara <-  as.data.frame(mean_sahara[order(mean_sahara$Mean, decreasing = T), ])

cit_usacan <- lapply(res_usacan, function(x) x$Countries)
long_cit_usacan <- data.frame(count = names(unlist(cit_usacan)), ncitations = unlist(cit_usacan))
mean_usacan <- 
  long_cit_usacan %>% 
  group_by(count) %>% 
  summarize(Mean = mean(ncitations, na.rm = T), std = sd(ncitations, na.rm = T))
cit_usacan <- as.data.frame(mean_usacan[order(mean_usacan$Mean, decreasing = T), ])

# joining data frames

all_cit <- rbind(cit_eastasia, cit_eu, cit_la, cit_middleafrica, cit_sahara, cit_usacan)

wb_reg <- 
rep(x = c("East Asia", "Europe", "Latin Am.", "Middle Africa", "Sahara", "USA-Can"), 
    times = c(dim(cit_eastasia)[1], dim(cit_eu)[1], dim(cit_la)[1], dim(cit_middleafrica)[1], dim(cit_sahara)[1], dim(cit_usacan)[1]))

all_cit$region <- wb_reg # adding region

```


## Processing data before plotting

Selecting only the top 10 countries that cite a given region and plotting the rarefaction values of citation for each country. These countries are the ones to be shown in Figure 1 of the manuscript

```{r}
# USA and Canada

names_top_usacan <- all_cit[which(all_cit$region == "USA-Can"), ][1:10, "count"]
count_top_usacan <- long_cit_usacan[which(is.na(match(long_cit_usacan$count, names_top_usacan)) != TRUE), ]
count_top_usacan_order <- count_top_usacan[order(count_top_usacan$ncitations, decreasing = T), ]
count_top_usacan_order$region.wb <- NA
count_top_usacan_order[count_top_usacan_order$count %in% toupper(names_LA), "region.wb"] <- "Latin America"
count_top_usacan_order[count_top_usacan_order$count %in% toupper(names_EastAsia), "region.wb"] <- "East Asia"
count_top_usacan_order[count_top_usacan_order$count %in% toupper(names_EU), "region.wb"] <- "Europe"
count_top_usacan_order[count_top_usacan_order$count %in% toupper(names_MidleEastAfrica), "region.wb"] <- "Middle Africa"
count_top_usacan_order[count_top_usacan_order$count %in% toupper(names_NA), "region.wb"] <- "USA and Canada"
count_top_usacan_order[count_top_usacan_order$count %in% toupper(names_SouthAsia), "region.wb"] <- "South Asia"
count_top_usacan_order[count_top_usacan_order$count %in% toupper(names_SubSahara), "region.wb"] <- "Sahara"

# East Asia
names_top_eastasia <- all_cit[which(all_cit$region == "East Asia"), ][1:10, "count"]
count_top_eastasia <- long_cit_eastasia[which(is.na(match(long_cit_eastasia$count, names_top_eastasia)) != TRUE), ]
count_top_eastasia_order <- count_top_eastasia[order(count_top_eastasia$ncitations, decreasing = T), ]
count_top_eastasia_order$region.wb <- NA
count_top_eastasia_order[count_top_eastasia_order$count %in% toupper(names_LA), "region.wb"] <- "Latin America"
count_top_eastasia_order[count_top_eastasia_order$count %in% toupper(names_EastAsia), "region.wb"] <- "East Asia"
count_top_eastasia_order[count_top_eastasia_order$count %in% toupper(names_EU), "region.wb"] <- "Europe"
count_top_eastasia_order[count_top_eastasia_order$count %in% toupper(names_MidleEastAfrica), "region.wb"] <- "Middle Africa"
count_top_eastasia_order[count_top_eastasia_order$count %in% toupper(names_NA), "region.wb"] <- "USA and Canada"
count_top_eastasia_order[count_top_eastasia_order$count %in% toupper(names_SouthAsia), "region.wb"] <- "South Asia"
count_top_eastasia_order[count_top_eastasia_order$count %in% toupper(names_SubSahara), "region.wb"] <- "Sahara"

# Latin America 
names_top_la <- all_cit[which(all_cit$region == "Latin Am."), ][1:10, "count"]
count_top_la <- long_cit_la[which(is.na(match(long_cit_la$count, names_top_la)) != TRUE), ]
count_top_la_order <- count_top_la[order(count_top_la$ncitations, decreasing = T), ]
count_top_la_order$region.wb <- NA
count_top_la_order[count_top_la_order$count %in% toupper(names_LA), "region.wb"] <- "Latin America"
count_top_la_order[count_top_la_order$count %in% toupper(names_EastAsia), "region.wb"] <- "East Asia"
count_top_la_order[count_top_la_order$count %in% toupper(names_EU), "region.wb"] <- "Europe"
count_top_la_order[count_top_la_order$count %in% toupper(names_MidleEastAfrica), "region.wb"] <- "Middle Africa"
count_top_la_order[count_top_la_order$count %in% toupper(names_NA), "region.wb"] <- "USA and Canada"
count_top_la_order[count_top_la_order$count %in% toupper(names_SouthAsia), "region.wb"] <- "South Asia"
count_top_la_order[count_top_la_order$count %in% toupper(names_SubSahara), "region.wb"] <- "Sahara"

# Europe
names_top_europe <- all_cit[which(all_cit$region == "Europe"), ][1:10, "count"]
count_top_europe <- long_cit_eu[which(is.na(match(long_cit_eu$count, names_top_europe)) != TRUE), ]
count_top_europe_order <- count_top_europe[order(count_top_europe$ncitations, decreasing = T), ]
count_top_europe_order$region.wb <- NA
count_top_europe_order[count_top_europe_order$count %in% toupper(names_LA), "region.wb"] <- "Latin America"
count_top_europe_order[count_top_europe_order$count %in% toupper(names_EastAsia), "region.wb"] <- "East Asia"
count_top_europe_order[count_top_europe_order$count %in% toupper(names_EU), "region.wb"] <- "Europe"
count_top_europe_order[count_top_europe_order$count %in% toupper(names_MidleEastAfrica), "region.wb"] <- "Middle Africa"
count_top_europe_order[count_top_europe_order$count %in% toupper(names_NA), "region.wb"] <- "USA and Canada"
count_top_europe_order[count_top_europe_order$count %in% toupper(names_SouthAsia), "region.wb"] <- "South Asia"
count_top_europe_order[count_top_europe_order$count %in% toupper(names_SubSahara), "region.wb"] <- "Sahara"

# Sahara

names_top_sahara <- all_cit[which(all_cit$region == "Sahara"), ][1:10, "count"]
count_top_sahara <- long_cit_sahara[which(is.na(match(long_cit_sahara$count, names_top_sahara)) != TRUE), ]
count_top_sahara_order <- count_top_sahara[order(count_top_sahara$ncitations, decreasing = T), ]
count_top_sahara_order$region.wb <- NA
count_top_sahara_order[count_top_sahara_order$count %in% toupper(names_LA), "region.wb"] <- "Latin America"
count_top_sahara_order[count_top_sahara_order$count %in% toupper(names_EastAsia), "region.wb"] <- "East Asia"
count_top_sahara_order[count_top_sahara_order$count %in% toupper(names_EU), "region.wb"] <- "Europe"
count_top_sahara_order[count_top_sahara_order$count %in% toupper(names_MidleEastAfrica), "region.wb"] <- "Middle Africa"
count_top_sahara_order[count_top_sahara_order$count %in% toupper(names_NA), "region.wb"] <- "USA and Canada"
count_top_sahara_order[count_top_sahara_order$count %in% toupper(names_SouthAsia), "region.wb"] <- "South Asia"
count_top_sahara_order[count_top_sahara_order$count %in% toupper(names_SubSahara), "region.wb"] <- "Sahara"

# Middle Africa

names_top_mafrica <- all_cit[which(all_cit$region == "Middle Africa"), ][1:10, "count"]
count_top_mafrica <- long_cit_middleafrica[which(is.na(match(long_cit_middleafrica$count, names_top_mafrica)) != TRUE), ]
count_top_mafrica_order <- count_top_mafrica[order(count_top_mafrica$ncitations, decreasing = T), ]
count_top_mafrica_order$region.wb <- NA
count_top_mafrica_order[count_top_mafrica_order$count %in% toupper(names_LA), "region.wb"] <- "Latin America"
count_top_mafrica_order[count_top_mafrica_order$count %in% toupper(names_EastAsia), "region.wb"] <- "East Asia"
count_top_mafrica_order[count_top_mafrica_order$count %in% toupper(names_EU), "region.wb"] <- "Europe"
count_top_mafrica_order[count_top_mafrica_order$count %in% toupper(names_MidleEastAfrica), "region.wb"] <- "Middle Africa"
count_top_mafrica_order[count_top_mafrica_order$count %in% toupper(names_NA), "region.wb"] <- "USA and Canada"
count_top_mafrica_order[count_top_mafrica_order$count %in% toupper(names_SouthAsia), "region.wb"] <- "South Asia"
count_top_mafrica_order[count_top_mafrica_order$count %in% toupper(names_SubSahara), "region.wb"] <- "Sahara"



```


# Rarefaction plots with citation values for each region

Assigning a color pallete to standardize the colors of the regions in Figure 1

```{r}
safe_colorblind_palette <- c("#88CCEE", "#CC6677", "#DDCC77", "#117733", "#332288", "#AA4499", 
                             "#44AA99", "#999933", "#882255", "#661100", "#6699CC", "#888888")

color_map <- c("Latin America" = "#117733", 
               "East Asia" = "#999933",
               "Europe" = "#88CCEE", 
               "Middle Africa" = "#AA4499",
               "USA and Canada" = "#661100", 
               "South Asia" = "#DDCC77",
               "Sahara" = "#888888")
```

Now we can plot the all the six violin plots with the mean values of citation that came from the rarefaction procedure 

```{r}
# violin for USA-Canada region

viol_usacan <- 
  count_top_usacan_order %>%
  mutate(ordcount = forcats::fct_reorder(count, ncitations)) %>% # Reorder data
  ggplot( aes(x=ordcount, y=ncitations, fill=region.wb)) +
  geom_violin(width=2.1, size=0.2) +
  scale_fill_manual(name = "Region", values = color_map) +
  labs(subtitle = "USA/Canada", title = "") +
  coord_flip() + # This switch X and Y axis and allows to get the horizontal version
  xlab("") +
  ylab("Citations") +
  annotate("text", x = 1.5, y = 250, 
           label = glue::glue("italic(J)=={round(J_usa, 3)}"), 
           parse = TRUE, size = 4) +
  theme_bw() +
  theme(panel.background = element_rect(fill = "transparent"),
        plot.margin = unit(c(0.1, 0.1, 0.1, 0.1), "mm"),
        legend.position = "none",
        legend.text = element_text(family = "Times", color = "black", size = 8),
        panel.border = element_blank())

# Violin for East Asia region

viol_eastasia <- 
  count_top_eastasia_order %>%
  mutate(ordcount = forcats::fct_reorder(count, ncitations)) %>% # Reorder data
  ggplot( aes(x=ordcount, y=ncitations, fill=region.wb)) +
  geom_violin(width=2.1, size=0.2) +
  scale_fill_manual(name = "Region", values = color_map) +
  labs(subtitle = "East Asia", title = "") +
  coord_flip() + # This switch X and Y axis and allows to get the horizontal version
  xlab("") +
  ylab("Citations") +
  annotate("text", x = 1.5, y = 200, 
           label = glue::glue("italic(J)=={round(J_east_asia, 3)}"), 
           parse = TRUE, size = 4) +
  theme_bw() +
  theme(panel.background = element_rect(fill = "transparent"),
        plot.margin = unit(c(0.1, 0.1, 0.1, 0.1), "mm"),
        legend.position = "none",
        legend.text = element_text(family = "Times", color = "black", size = 8),
        panel.border = element_blank())

# Violin for Latin America region

viol_la <- 
  count_top_la_order %>%
  mutate(ordcount = forcats::fct_reorder(count, ncitations)) %>% # Reorder data
  ggplot( aes(x=ordcount, y=ncitations, fill=region.wb)) +
  geom_violin(width=2.1, size=0.2) +
  scale_fill_manual(name = "Region", values = color_map) +
  labs(subtitle = "Latin America", title = "") +
  coord_flip() + # This switch X and Y axis and allows to get the horizontal version
  xlab("") +
  ylab("Citations") +
  annotate("text", x = 1.5, y = 200, 
           label = glue::glue("italic(J)=={round(J_la, 3)}"), 
           parse = TRUE, size = 4) +
  theme_bw() +
  theme(panel.background = element_rect(fill = "transparent"),
        plot.margin = unit(c(0.1, 0.1, 0.1, 0.1), "mm"),
        legend.position = "none",
        legend.text = element_text(family = "Times", color = "black", size = 10),
        panel.border = element_blank())

# Violin plot for Europe region

viol_europe <- 
  count_top_europe_order %>%
  mutate(ordcount = forcats::fct_reorder(count, ncitations)) %>% # Reorder data
  ggplot( aes(x=ordcount, y=ncitations, fill=region.wb)) +
  geom_violin(width=2.1, size=0.2) +
  scale_fill_manual(name = "Region", values = color_map) +
  labs(subtitle = "Europe", title = "") +
  coord_flip() + # This switch X and Y axis and allows to get the horizontal version
  xlab("") +
  ylab("Citations") +
  annotate("text", x = 1.5, y = 200, 
           label = glue::glue("italic(J)=={round(J_eu, 3)}"), 
           parse = TRUE, size = 4) +
  theme_bw() +
  theme(panel.background = element_rect(fill = "transparent"),
        plot.margin = unit(c(0.1, 0.1, 0.1, 0.1), "mm"),
        legend.position = "none",
        axis.text.x = element_text(family = "Times", color = "black", size = 9),
        legend.text = element_text(family = "Times", color = "black", size = 8),
        panel.border = element_blank())

# Plot for Su-Saharan region

viol_sahara <- 
  count_top_sahara_order %>%
  mutate(ordcount = forcats::fct_reorder(count, ncitations)) %>% # Reorder data
  ggplot( aes(x=ordcount, y=ncitations, fill=region.wb)) +
  geom_violin(width=2.1, size=0.2) +
  scale_fill_manual(name = "Region", values = color_map) +
  labs(subtitle = "Sub-Saharan", title = "") +
  coord_flip() + # This switch X and Y axis and allows to get the horizontal version
  xlab("") +
  ylab("Citations") +
  scale_y_continuous(breaks = c(seq(from = 50, to = 250, by = 75))) +
  annotate("text", x = 1.5, y = 170, 
           label = glue::glue("italic(J)=={round(J_sahara, 3)}"), 
           parse = TRUE, size = 4) +
  theme_bw() +
  theme(panel.background = element_rect(fill = "transparent"),
        plot.margin = unit(c(0.1, 0.1, 0.1, 0.1), "mm"),
        axis.text.x = element_text(family = "Times", color = "black", size = 7.5),
        legend.position = "none",
        legend.text = element_text(family = "Times", color = "black", size = 8),
        panel.border = element_blank())
  
# Violin for Middle Africa region

viol_middleafrica <- 
  count_top_mafrica_order %>%
  mutate(ordcount = forcats::fct_reorder(count, ncitations)) %>% # Reorder data
  ggplot( aes(x=ordcount, y=ncitations, fill=region.wb)) +
  geom_violin(width=2.1, size=0.2) +
  scale_fill_manual(name = "Region", values = color_map) +
  labs(subtitle = "Middle Africa", title = "") +
  coord_flip() + # This switch X and Y axis and allows to get the horizontal version
  xlab("") +
  ylab("Citations") +
  annotate("text", x = 1.5, y = 200, 
           label = glue::glue("italic(J)=={round(J_mafrica, 3)}"), 
           parse = TRUE, size = 4) +
  theme_bw() +
  theme(panel.background = element_rect(fill = "transparent"),
        plot.margin = unit(c(0.1, 0.1, 0.1, 0.1), "mm"),
        legend.position = "none",
        legend.text = element_text(family = "Times", color = "black", size = 8),
        panel.border = element_blank())
  

```


Joining all violin plots

```{r}
legend <- cowplot::get_legend(viol_la + theme(legend.position = "bottom")) 
all_violin <- 
  cowplot::plot_grid(viol_la, viol_usacan, viol_eastasia, viol_europe, viol_sahara, viol_middleafrica, ncol = 3)
plot_violin <- cowplot::plot_grid(all_violin, legend, ncol = 1, rel_heights = c(2, 0.5))

```
