==> R/03_S_figure1-main-text.Rmd <==
---
title: "Figure main text"
author: "Gabriel Nakamura"
date: "2023-07-24"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

Joining plots - citation and geo markers

```{r}
library(patchwork)
plot_cit_geoMark <- 
  map_markers_wb / plot_violin  + plot_annotation(tag_levels = "a")

```


# Saving plots
```{r}
ggplot2::ggsave(filename = here::here("output", "map_geo_markers_wb.png"), plot = map_markers_wb, device = "png", dpi = 200, width = 10, height = 8)

ggplot2::ggsave(filename = here::here("output", "all_violin.png"), plot = all_violin, device = "png", dpi = 200, width = 10, height = 8)

ggplot2::ggsave(filename = here::here("output", "Fig1_cit_geoMark.png"), plot = plot_cit_geoMark, device = "png", dpi = 300, width = 15, height = 12)


```
